Extract comments to component
